backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
passlib[bcrypt]
pyjwt
pydantic
langchain[google-genai]
